backend/requirements.txt
flask==2.3.0
flask-cors==4.0.0
tensorflow==2.15.0
numpy==1.24.3
pillow==10.0.0